src/test/java/com/teragrep/snw_01/TableOffsetTest.java
package com.teragrep.snw_01;

/*
 * Teragrep ServiceNow Integrator SNW-01
 * Copyright (C) 2023 Suomen Kanuuna Oy
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://github.com/teragrep/teragrep/blob/main/LICENSE>.
 *
 *
 * Additional permission under GNU Affero General Public License version 3
 * section 7
 *
 * If you modify this Program, or any covered work, by linking or combining it
 * with other code, such other code is not for that reason alone subject to any
 * of the requirements of the GNU Affero GPL version 3 as long as this Program
 * is the same Program as licensed from Suomen Kanuuna Oy without any additional
 * modifications.
 *
 * Supplemented terms under GNU Affero General Public License version 3
 * section 7
 *
 * Origin of the software must be attributed to Suomen Kanuuna Oy. Any modified
 * versions must be marked as "Modified version of" The Program.
 *
 * Names of the licensors and authors may not be used for publicity purposes.
 *
 * No rights are granted for use of trade names, trademarks, or service marks
 * which are in The Program if any.
 *
 * Licensee must indemnify licensors and authors for any liability that these
 * contractual assumptions impose on licensors and authors.
 *
 * To the extent this program is licensed as part of the Commercial versions of
 * Teragrep, the applicable Commercial License may apply to this file if you as
 * a licensee so wish it.
 */

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.time.Instant;
import java.time.ZoneId;
import java.time.ZonedDateTime;

public class TableOffsetTest {

    @Test
    public void testTableOffset() {

        System.setProperty("statestore.directory", "target");

        // Sun Oct 22 09:40:00 PM EEST 2023
        Instant instant = Instant.ofEpochSecond(1698000000);
        ZonedDateTime startOf = ZonedDateTime.ofInstant(instant, ZoneId.of("Europe/Helsinki"));

        // initial
        TableDay tableDay = new TableDay(startOf, 5, "foo");
        TableOffset tableOffset = new TableOffset(tableDay);

        Assertions.assertEquals(0, tableOffset.getOffset());
        tableOffset.increment(5);
        Assertions.assertEquals(5, tableOffset.getOffset());
        tableOffset.increment(5);
        Assertions.assertEquals(10, tableOffset.getOffset());

        // test persistence
        TableOffset sameDayOffset = new TableOffset(tableDay);
        Assertions.assertEquals(10, sameDayOffset.getOffset());

    }
}
